refactor: split tracker code into separate files
